refactor finalize helpers
